json format - finalised to run for distributed setup

+++ src/com/project2/custominputformat/CustomInputFormat.java
@@ -1,13 +1,12 @@
 /**
-@created on: 18/3/19,
-@author: Shreesha N,
-@version: v0.0.1
-@system name: badgod
-Description:
-
-..todo::
-
-*/
+ * @created on: 18/3/19,
+ * @author: Shreesha N,
+ * @version: v0.0.1
+ * @system name: badgod
+ * Description:
+ * <p>
+ * ..todo::
+ */
 
 package com.project2.custominputformat;
 
@@ -73,15 +72,17 @@ public class CustomInputFormat {
 
         Configuration conf = new Configuration();
 
+        String inputDataPath = args[0];
+        String outputPath = args[1];
+
         // add the below code if you are reading/writing from/to HDFS
-//        String inputDataPath = "hdfs://localhost:9000/ds503/hw2/input/airfield.json";
-//        String outputPath = "hdfs://localhost:9000/ds503/hw2/output/custominputformat/";
+        String hadoopHome = System.getenv("HADOOP_HOME");
+        if (hadoopHome == null) {
+            throw new Exception("HADOOP_HOME not found. Please make sure system path has HADOOP_HOME point to hadoop installation directory");
+        }
+        conf.addResource(new Path(hadoopHome + "/etc/hadoop/core-site.xml"));
+        conf.addResource(new Path(hadoopHome + "/etc/hadoop/hdfs-site.xml"));
 
-        String inputDataPath = "/Users/badgod/badgod_documents/github/BigDataTutorials/input/project2/airfield.json";
-        String outputPath = "/Users/badgod/badgod_documents/github/BigDataTutorials/output/project2/custominputformat/";
-
-//        conf.addResource(new Path("/Users/badgod/badgod_documents/technologies/hadoop-3.1.2/etc/hadoop/core-site.xml"));
-//        conf.addResource(new Path("/Users/badgod/badgod_documents/technologies/hadoop-3.1.2/etc/hadoop/hdfs-site.xml"));
         FileSystem fs = FileSystem.get(conf);
         fs.delete(new Path(outputPath), true);
         Job job = Job.getInstance(conf, "CustomInputFormat");
@@ -99,5 +100,7 @@ public class CustomInputFormat {
         FileOutputFormat.setOutputPath(job, new Path(outputPath));
 
         job.waitForCompletion(true);
+        if (fs != null)
+            fs.close();
     }
 }
